+++ CHANGELOG.md
@@ -81,6 +81,7 @@ SPDX-License-Identifier: CC-BY-SA-4.0
 - The email sign-in/registration does not require an email address anymore but uses a username
 - The history shows both the entries saved in LocalStorage and the entries saved on the server together
 - The gist embedding uses a click-shield, like vimeo and youtube
+- HTML-Iframes are capsuled in click-shields
 - Use [Twemoji](https://twemoji.twitter.com/) as icon font
 - The `[name=...]`, `[time=...]` and `[color=...]` tags may now be used anywhere in the document and not just inside of blockquotes and lists.
 - The <i class="fa fa-picture-o"/> (add image) and <i class="fa fa-link"/> (add link) toolbar buttons put selected links directly in the `()` instead of the `[]` part of the generated markdown.